fix scrolling on bounties and organize components
