+++ pages/[domain]/bounties/index.tsx
@@ -1,5 +1,4 @@
-import Box from '@mui/material/Box';
-import { BountyList } from 'components/bounties/BountyList';
+import BountyList from 'components/bounties/BountyList';
 import PageLayout from 'components/common/PageLayout';
 import { setTitle } from 'hooks/usePageTitle';
 import { ReactElement } from 'react';
@@ -9,9 +8,7 @@ export default function BountyPage () {
   setTitle('Bounties');
 
   return (
-    <Box py={3} px='80px'>
-      <BountyList />
-    </Box>
+    <BountyList />
   );
 
 }
